Updated server to only show you the tasks that you own. Website has been edited to comply with these changes.

+++ Flex-Server/server.py
@@ -21,13 +21,15 @@ class Task(db.Model):
     mAssignee = db.Column(db.String())
     mDueDate = db.Column(db.String())
     mTaskStatus = db.Column(db.String())
+    mOwner = db.Column(db.String())
 
-    def __init__(self, mTask, mDescription, mAssignee, mDueDate, mTaskStatus):
+    def __init__(self, mTask, mDescription, mAssignee, mDueDate, mTaskStatus, mOwner):
         self.mTask = mTask
         self.mDescription = mDescription
         self.mAssignee = mAssignee
         self.mDueDate = mDueDate
         self.mTaskStatus = mTaskStatus
+        self.mOwner = mOwner
 
     def __repr__(self):
         return '<Task %r>' % self.mID
@@ -141,13 +143,30 @@ def app_task():
         mDescription = data.get('mDescription')
         mAssignee = data.get('mAssignee')
         mDueDate = data.get('mDueDate')
-        print(mTask + " " + mDescription + " " + mAssignee + " " + mDueDate)
-        db.session.add(Task(mTask, mDescription, mAssignee, mDueDate, "To Do"))
-        db.session.commit()
-        response["response"] = "Added The Task!"
+        uuid = data.get('uuid')
+        mOwner = data.get('mOwner')  # Either "personal" or "team", if "team" then the task will be owned by the team
+        # and you have to include 'mTeam'
+        if mOwner == "personal":
+            mOwner = uuid
+            db.session.add(Task(mTask, mDescription, mAssignee, mDueDate, "To Do", mOwner))
+            db.session.commit()
+            response["response"] = "added personal task"
+            response["status"] = "success"
+            return jsonify(**response)
+        elif mOwner == "team":
+            try:
+                mTeam = data.get('mTeam')  # The Team's ID
+                mOwner = mTeam
+                db.session.add(Task(mTask, mDescription, mAssignee, mDueDate, "To Do", mOwner))
+                db.session.commit()
+                response["response"] = "Added the task!"
+                return jsonify(**response)
+            except:
+                response["response"] = "missing"
+                return jsonify(**response)
     except:
-        response["response"] = "failed"
-    return jsonify(**response)
+        response["response"] = "missing"
+        return jsonify(**response)
 
 
 @app.route('/upgradeTask', methods=['POST'])
@@ -156,47 +175,56 @@ def upgrade_task():
     response = {}
     try:
         mID = data.get('mID')
-        print(mID)
+        uuid = data.get('uuid')
         task = Task.query.filter(Task.mID == mID).first()
-        if task.mTaskStatus == "To Do":
-            task.mTaskStatus = "In Process"
-        elif task.mTaskStatus == "In Process":
-            task.mTaskStatus = "Done"
-        db.session.commit()
-        response["response"] = "Successfully upgraded task"
+        if task.mOwner == uuid:
+            if task.mTaskStatus == "To Do":
+                task.mTaskStatus = "In Process"
+            elif task.mTaskStatus == "In Process":
+                task.mTaskStatus = "Done"
+            db.session.commit()
+            response["response"] = "Successfully upgraded task"
+        else:
+            response["status"] = "error"
+            response["response"] = "You don't own this task!"
     except:
         response["response"] = "failed to upgrade task"
     return jsonify(**response)
 
 
-@app.route('/getTasks')
+@app.route('/getTasks', methods=['POST'])
 def get_tasks():
-    isInitTodo = False
-    isInitDoing = False
-    isInitDone = False
-    tasks = []
-    for task in Task.query.all():
-        if task.mTaskStatus == "To Do":
-            isInitTodo = True
-        elif task.mTaskStatus == "In Process":
-            isInitDoing = True
-        elif task.mTaskStatus == "Done":
-            isInitDone = True
-        response = {'mTask': task.mTask,
-                    'mDescription': task.mDescription,
-                    'mAssignee': task.mAssignee,
-                    'mDueDate': task.mDueDate,
-                    'mID': task.mID,
-                    'mTaskStatus': task.mTaskStatus
-                    }
-        tasks.append(response)
+    data = request.json
     response = {}
-    response['tasks'] = tasks
-    response['meta'] = {}
-    response['meta']['isInitTodo'] = isInitTodo
-    response['meta']['isInitDoing'] = isInitDoing
-    response['meta']['isInitDone'] = isInitDone
-    return jsonify(**response)
+
+    try:
+        uuid = data.get('uuid')
+        scope = data.get('scope')
+    except:
+        response["response"] = "missing"
+        return jsonify(**response)
+
+    if check_if_valid_session(uuid) is False:
+        response["response"] = "invalid uuid"
+        return jsonify(**response)
+    else:
+        if scope == "personal":
+            tasks = []
+            for task in Task.query.filter(Task.mOwner == uuid).all():
+                taskInfo = {'mTask': task.mTask,
+                            'mDescription': task.mDescription,
+                            'mAssignee': task.mAssignee,
+                            'mDueDate': task.mDueDate,
+                            'mID': task.mID,
+                            'mTaskStatus': task.mTaskStatus
+                            }
+                tasks.append(taskInfo)
+            response['tasks'] = tasks
+            response['meta'] = {}
+            response["response"] = "success"
+        else:
+            response["response"] = "not implemented yet"
+        return jsonify(**response)
 
 
 @app.route('/deleteTask', methods=['POST'])
@@ -206,11 +234,19 @@ def delete_tank():
 
     try:
         mID = data.get('mID')
-        Task.query.filter(Task.mID == mID).delete()
-        db.session.commit()
-        response["response"] = "deleted task"
+        uuid = data.get('uuid')
+        task = Task.query.filter(Task.mID == mID).first()
+        if task is not None and task.mOwner == uuid:
+            Task.query.filter(Task.mID == mID).delete()
+            db.session.commit()
+            response["response"] = "deleted task"
+            response["status"] = "success"
+        else:
+            response["response"] = "You don't own this task or this task does not exist!"
+            response["status"] = "error"
     except:
         response["response"] = "failed at deleting task"
+        response["status"] = "error"
     return jsonify(**response)
 
 
@@ -337,18 +373,25 @@ def update_task():
         response["response"] = "missing"
         return jsonify(**response)
 
-    try:
-        task = Task.query.filter(Task.mID == mId).first()
-        task.mDescription = mDescription
-        task.mTask = mTitle
-        task.mAssignee = mAssignee
-        task.mDueDate = mDueDate
-        db.session.commit()
-        response["response"] = "success"
-        return jsonify(**response)
-    except:
-        response["response"] = "error"
-        return jsonify(**response)
+    task = Task.query.filter(Task.mID == mId).first()
+
+    if task is not None and task.mOwner == uuid:
+        try:
+            task.mDescription = mDescription
+            task.mTask = mTitle
+            task.mAssignee = mAssignee
+            task.mDueDate = mDueDate
+            db.session.commit()
+            response["response"] = "success"
+            response["status"] = "success"
+        except:
+            response["response"] = "error"
+            response["status"] = "error"
+    else:
+        response["response"] = "You don't own this task or this task does not exist"
+        response["status"] = "error"
+
+    return jsonify(**response)
 
 
 @app.route('/joinTeam', methods=['POST'])
